refactor(ui): Sync-Fehler-Funktionen + Formatier-Helfer nach sync_orchestrator (#49)
